Rename Project 3 Trainity.pdf to Operation Analytics and Investigating Metric Spike.pdf
Operation Analytics is the analysis done for the complete end to end operation of company.  with the help of this and Investigating the metric spike company finds the area on which it 
must improve upon. this is the most important task of every company that helps them to predict the 
overall growth or decline of company fortunes.