Add files via upload
The aim of this project is to identify and analyzing the factor to identify the difficulty faced by customer during their installment. With the EDA(Exploratory data analysis) bank will understand the patterns in loan application data and they will ensure that applicant not rejected with the help of given dataset.