Add files via upload
This project deal with the car industry and its featutres that will impact        the price and profitability of car in future. The automated industry has been evolving over the past few decade with the growing fuel efficiency and technologies. Nowadays electric and hybrid vehicle are gaining focus in market, with the various variety it is become important to understand the factor that affecting the deamand of cars. By using the data analysis technique like regression analysis and market segmentation, the manufacturer could develop a strategy that balance demand with profitability.